orchestrator/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.2
watchdog==6.0.0
openai==0.27.0
Flask==3.0.0
Flask-CORS==4.0.0
Jinja2==3.1.3
Werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
blinker==1.7.0
